backend/requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.6.1
Django==2.2.1
django-rest-knox==4.0.1
djangorestframework==3.9.4
entrypoints==0.3
flake8==3.7.7
mccabe==0.6.1
psycopg2==2.8.2
psycopg2-binary==2.8.2
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
python-decouple==3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
